@@ -2,6 +2,8 @@ package org.datastax.vsdemo.indexing;
 
 import org.datastax.vsdemo.indexing.messages.IndexRequest;
 import org.datastax.vsdemo.indexing.messages.SimilarityRequest;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
@@ -14,6 +16,8 @@ import java.util.List;
 public class IndexingController {
     private final IndexingService service;
     private final SimpMessagingTemplate simp;
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(IndexingController.class);
 
     public IndexingController(IndexingService service, SimpMessagingTemplate simp) {
         this.service = service;
@@ -36,7 +40,7 @@ public class IndexingController {
         service.getSimilarSentences(userID, message.query(), message.limit())
             .whenComplete((result, error) -> {
                 if (error != null) {
-                    error.printStackTrace();
+                    LOGGER.error("Error in getting similar sentences", error);
                 } else {
                     simp.convertAndSend("/topic/query-result", result);
                 }
