@@ -1,12 +1,14 @@
-package org.datastax.vsdemo.indexing;
+package org.datastax.vsdemo.indexing.utils;
 
 import com.datastax.oss.driver.api.core.data.CqlVector;
 import com.datastax.oss.driver.shaded.guava.common.collect.ImmutableList;
 
 import java.util.List;
 
+import static org.datastax.vsdemo.indexing.utils.Prelude.map;
+
 public class VectorUtils {
-    public static double testSimilarity(CqlVector<Float> vecA, CqlVector<Float> vecB) {
+    public static double cosineSimilarity(CqlVector<Float> vecA, CqlVector<Float> vecB) {
         var listA = (ImmutableList<Float>) vecA.getValues();
         var listB = (ImmutableList<Float>) vecB.getValues();
 
@@ -24,8 +26,12 @@ public class VectorUtils {
         return Math.max(similarity, 0);
     }
 
+    public static double colbertSimilarity(CqlVector<Float> vecA, List<CqlVector<Float>> vecB) {
+        return 0.0;
+    }
+
     @SuppressWarnings("unchecked")
-    public static CqlVector<Float> listToCqlVector(List<Float> list) {
-        return CqlVector.builder().addAll(list).build();
+    public static CqlVector<Float> listToCqlVector(List<Double> list) {
+        return CqlVector.builder().addAll(map(Double::floatValue, list)).build();
     }
 }
